Split client into multiple files and add more tests
