+++ woodpecker-go/woodpecker/client_test.go
@@ -25,44 +25,6 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func Test_QueueInfo(t *testing.T) {
-	fixtureHandler := func(w http.ResponseWriter, _ *http.Request) {
-		fmt.Fprint(w, `{
-			"pending": null,
-			"running": [
-					{
-							"id": "4696",
-							"data": "",
-							"labels": {
-									"platform": "linux/amd64",
-									"repo": "woodpecker-ci/woodpecker"
-							},
-							"Dependencies": [],
-							"DepStatus": {},
-							"RunOn": null
-					}
-			],
-			"stats": {
-					"worker_count": 3,
-					"pending_count": 0,
-					"waiting_on_deps_count": 0,
-					"running_count": 1,
-					"completed_count": 0
-			},
-			"Paused": false
-	}`)
-	}
-
-	ts := httptest.NewServer(http.HandlerFunc(fixtureHandler))
-	defer ts.Close()
-
-	client := NewClient(ts.URL, http.DefaultClient)
-
-	info, err := client.QueueInfo()
-	assert.NoError(t, err)
-	assert.Equal(t, 3, info.Stats.Workers)
-}
-
 func Test_LogLevel(t *testing.T) {
 	logLevel := "warn"
 	fixtureHandler := func(w http.ResponseWriter, r *http.Request) {
